Rename kotlin-console to kotlin-cli
