+++ samples/kotlin-cli/src/main/kotlin/org/jellyfin/sample/cli/Main.kt
@@ -1,14 +1,14 @@
-package org.jellyfin.sample.console
+package org.jellyfin.sample.cli
 
 import kotlinx.cli.ArgParser
 import org.jellyfin.apiclient.Jellyfin
 import org.jellyfin.apiclient.model.ClientInfo
 import org.jellyfin.apiclient.model.DeviceInfo
-import org.jellyfin.sample.console.cli.*
+import org.jellyfin.sample.cli.command.*
 
 fun main(args: Array<String>) {
 	val jellyfin = Jellyfin {
-		clientInfo = ClientInfo("Jellyfin Sample: Kotlin Console", "DEV")
+		clientInfo = ClientInfo("Jellyfin Sample: Kotlin CLI", "DEV")
 		deviceInfo = DeviceInfo("cli", "cli")
 	}
